# data-ref/src/main/resources/META-INF/resources/openapi.yaml
openapi: 3.0.3
info:
  title: API Baseline
  description: CRUD Operation on repositories data
  version: 0.7.0
servers:
  - url: http://data-ref.provoly.com
paths:
...
  /metadata:
    post:
      tags:
        - Metadata Def Controller
      operationId: com.provoly.ref.metadata.MetadataDefController_addMetadata
      summary: Save a new metadata definition
      description: |
        A metadata is used to add context to an entity. 
        Metadata definitions must first be created before being used on an entity.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MetadataDefDto'
            examples:
              addMetadataDef:
                summary: Add a valid metadata definition
                value: |
                  {
                     "id": "ba681f33-fe70-42f8-a277-05c607ba29cf",
                     "name": "metadata_test",
                     "type": "STRING"
                  }
              addInvliadMetadataDef:
                summary: Add a metadata definition with invalid type
                value: |
                  {
                     "id": "ba681f33-fe70-42f8-a277-05c607ba29cf",
                     "name": "metadata_test",
                     "type": "UNKNOWN"
                  }
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          content:
            application/json:
              examples:
                nameAlreadyExists:
                  summary: Add metadata definition with a name that already exists
                  value: |
                    {
                      "application": "data-ref",
                      "code": "NAME_ALREADY_USED",
                      "message": "Name {{metadata_def_name}} already exists"
                    }
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
        - oidc:
            - metadata_item_ref_write
    get:
      tags:
        - Metadata Def Controller
      summary: Retrieve all metadata definitions
      operationId: com.provoly.ref.metadata.MetadataDefController_getAll
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                uniqueItems: true
                type: array
                items:
                  $ref: '#/components/schemas/MetadataDefDto'
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
        - oidc:
            - metadata_item_ref_read
            - item_write
            - dashboard_write
            - dataset_write
  /metadata/id/{metadataId}:
    get:
      tags:
        - Metadata Def Controller
      operationId: com.provoly.ref.metadata.MetadataDefController_getById
      summary: Retrieve a metadata definition based on its identifier
      parameters:
        - name: metadataId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetadataDefDto'
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "404":
          $ref: '#/components/responses/notFoundException'
      security:
        - oidc:
            - metadata_item_ref_read
            - item_write
            - search
            - dashboard_read
  /metadata/name/{metadataName}:
    get:
      tags:
        - Metadata Def Controller
      operationId: com.provoly.ref.metadata.MetadataDefController_getByName
      summary: Retrieve a metadata based on its name
      parameters:
        - name: metadataName
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetadataDefDto'
        "404":
          description: Bad Request
          content:
            application/json:
              examples:
                notFound:
                  summary: Not found
                  value: |
                    {
                      "application": "data-ref",
                      "code": "NOT_FOUND",
                      "message": "Unknown metadata name : {{metadata_name}}"
                    }
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
        - oidc:
            - metadata_item_ref_read
            - item_write
            - search
            - datasource_read
  /metadata/slug/{metadataSlug}:
    get:
      tags:
        - Metadata Def Controller
      operationId: com.provoly.ref.metadata.MetadataDefController_getBySlug
      summary: Retrieve a metadata based on its slug
      parameters:
        - name: metadataSlug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetadataDefDto'
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "404":
          description: Bad Request
          content:
            application/json:
              examples:
                notFound:
                  summary: Not found
                  value: |
                    {
                      "application": "data-ref",
                      "code": "NOT_FOUND",
                      "message": "Unknown metadata with slug : {{metadata_slug}}"
                    }
      security:
        - oidc:
            - metadata_item_ref_read
            - item_write
            - search
            - datasource_read
...
  /links:
    get:
      tags:
        - Link Controller
      summary: Retrieve all links
      description: |
        A link is a connections or relationships between attributes of classes. 
        They allow us to represent associations between our objects.
      operationId: com.provoly.ref.link.LinkController_getAll
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LinkDetailsDto'
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "404":
          $ref: '#/components/responses/notFoundException'
      security:
        - oidc:
            - link_read
    post:
      tags:
        - Link Controller
      summary: Save a new link defining a relationship
      operationId: com.provoly.ref.link.LinkController_saveLink
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LinkDto'
            examples:
              add_link:
                summary: Adding a link defining a relationship
                value: |
                  {
                      "id": "334c1229-61e7-4807-aeaa-edaf16ec2273",
                      "relationType": "2f021992-5fb0-42d3-9646-0affccb98105",
                      "attributeSource": "f897dc76-9c4a-4b7a-9797-fc577c0438e9",
                      "attributeDestination": "7319d99b-5400-40e4-a0e5-f9fb46b3d2d6"
                  }
      responses:
        "201":
          description: Created
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "409":
          description: Conflict
          content:
            application/json:
              examples:
                alreadyExists:
                  summary: Link already exist
                  description: Occurs when you try to create a link that already exists
                  value: |
                    {
                      "application": "data-ref",
                      "code": "CONFLICT",
                      "message": "Link from {{attributeSource}} to {{attributeDestination}} already exists"
                    }
      security:
        - oidc:
            - link_write
  /links/id/{id}:
    get:
      tags:
        - Link Controller
      summary: Retrieve a link based on its identifier
      operationId: com.provoly.ref.link.LinkController_getById
      parameters:
        - name: id
          in: path
          description: Link UUID
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkDetailsDto'
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "404":
          $ref: '#/components/responses/notFoundException'
      security:
        - oidc:
            - link_read
            - update_relation_aggregate
    delete:
      tags:
        - Link Controller
      summary: delete a link based on its identifier
      operationId: com.provoly.ref.link.LinkController_deleteLink
      parameters:
        - name: id
          in: path
          description: Link UUID
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        "204":
          description: No Content
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "404":
          $ref: '#/components/responses/notFoundException'
      security:
        - oidc:
            - link_write
...
  /notification:
    post:
      tags:
        - Notification Controller
      summary: Save a new notification to the associated users
      operationId: com.provoly.ref.message.notification.NotificationController_addNotification
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationRequestDto'
            examples:
              default:
                summary: Adding a notification to the associated users
                value: |
                  {
                    "type": "notification",
                    "id": "c5b4908f-c477-436b-b346-160803a82915",
                    "notificationTextDto": {
                      "code": "DASHBOARD_PUBLIC",
                      "param": {
                        "name": "presentationAdmin"
                      },
                      "link": null
                    },
                    "users": ["7b2c2947-d26d-41c8-8827-23b43b6ff5b3", "941cc554-b81d-4998-a65d-8f7861ab110d", "8c002d8a-a050-40e3-bf28-a6217bf8f6dc"]
                  }
      responses:
        "201":
          description: Created
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
        - oidc:
            - notification_write
...
  /notification/me/id/{id}:
    delete:
      tags:
        - Notification Controller
      summary: Remove a notification from its associated user
      operationId: com.provoly.ref.message.notification.NotificationController_acknowledgeNotification
      parameters:
        - name: id
          in: path
          description: Notification UUID
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        "204":
          description: No Content
        "404":
          $ref: '#/components/responses/notFoundException'
components:
  responses:
    notFoundException:
      description: Entity not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
          example: 404
        message:
          type: string
          example: "Entity : ac2c1a7f-0694-4f03-bb00-d349adeecd3d inexistant."
      required:
        - code
        - message
...
    AttributeDefDetailsDto:
      type: object
      properties:
        oclass:
          $ref: '#/components/schemas/UUID'
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
          example: "Attribute name"
        technicalName:
          type: string
          example: "Technical name"
        slug:
          type: string
          example: "12390_attribute_name"
        multiValued:
          type: boolean
          example: false
        field:
          $ref: '#/components/schemas/FieldDto'
...
    FieldDto:
      type: object
      required:
        - id
        - name
        - type
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
          example: "my field"
        type:
          $ref: '#/components/schemas/Type'
        slug:
          type: string
          example: "73981_my_field"
        crs:
          type: string
          example: "EPSG:1233"
          default: "EPSG:4326"
          description: "Required only for Geo field. Should be a valid CRS"
...
    LinkDetailsDto:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        relationType:
          $ref: '#/components/schemas/RelationTypeDto'
        attributeSource:
          $ref: '#/components/schemas/AttributeDefDetailsDto'
        attributeDestination:
          $ref: '#/components/schemas/AttributeDefDetailsDto'
    LinkDto:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        relationType:
          $ref: '#/components/schemas/UUID'
        attributeSource:
          $ref: '#/components/schemas/UUID'
        attributeDestination:
          $ref: '#/components/schemas/UUID'
...
    MetadataDefDto:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
          example: "my metadata definition"
        type:
          $ref: '#/components/schemas/VariableTypeDto'
        description:
          type: string
          example: "metadata definition description"
        allowedValues:
          type: array
          items:
            type: string
        slug:
          type: string
          example: "87280_my_metadata_definition"
        readOnly:
          type: boolean
        system:
          type: boolean
...
    NotificationMessageCode:
      enum:
        - DASHBOARD_PRIVATE
        - DASHBOARD_PUBLIC
        - DASHBOARD_DELETED
      type: string
    NotificationRequestDto:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        users:
          type: array
          items:
            format: uuid
            pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
            type: string
        notificationTextDto:
          $ref: '#/components/schemas/NotificationTextDto'
        link:
          type: string
    NotificationTextDto:
      type: object
      properties:
        code:
          $ref: '#/components/schemas/NotificationMessageCode'
        param:
          type: object
          additionalProperties:
            type: string
...
    RelationTypeDto:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
        slug:
          type: string
...
    Type:
      enum:
        - string
        - keyword
        - integer
        - long
        - decimal
        - raw
        - instant
        - Point
        - MultiPoint
        - LineString
        - MultiLineString
        - Polygon
        - MultiPolygon
      type: string
      example: "keyword"
...
    UUID:
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
      type: string
...
    VariableTypeDto:
      enum:
        - INTEGER
        - STRING
        - DOUBLE
        - DATE
        - UUID
        - LIST
      type: string
...
  securitySchemes:
    oidc:
      type: openIdConnect
      description: Authentication
      openIdConnectUrl: https://auth/realms/provoly/.well-known/openid-configuration